Fix handling of empty superscripts and subscripts
Temml raised an exception when I fed it an equation containing an empty
superscript, e.g. $x^{}$. Its a weird but valid bit of LaTeX math.

The problem arose in some code that's directed at working around a
Chromium bug in temml.cjs:7634. It attempts to look at the first child
in a superscript to see if it needs to apply the fix. But if the
superscript is empty, there is no first element. I added a check
for an empty superscript here